Updates requested

+++ database/migrations/2020_06_20_115811_create_scheduled_messages_table.php
@@ -25,7 +25,7 @@ class CreateScheduledMessagesTable extends Migration
             $table->string('name');
             $table->text('message');
 
-            $table->tinyInteger('time');
+            $table->unsignedInteger('time');
 
             $table->timestamps();
         });
